Improve pane title
The pane originally named "Canvas the file embedded" (meaning canvases that are embedded in the file) does not correctly describe the functionality of the pane, which is to display canvases that embed the file. I propose to rename this pane to "Canvases embedding the file" to better reflect what the pane does.